requirements-functions.txt
functions-framework==3.5.0
requests==2.31.0
google-generativeai==0.3.2
google-cloud-storage==2.10.0
